Rename file containing app entrypoint
